feat(process): widen X registers from 256 to 1024 (BEAM spec), bump v0.3.5
BEAM supports up to 1024 X registers. OTP 27 compiled Gleam code can use
indices beyond 255 for complex pattern matches. Widens the X register
array from [Term; 256] to [Term; 1024] and the index type from u8 to u16
throughout the interpreter.

+++ crates/beamr/src/interpreter/opcodes/trampoline.rs
@@ -70,7 +70,7 @@ pub fn handle_trampoline(
 
     // Load arguments into x registers.
     for (index, arg) in trampoline.args.iter().enumerate() {
-        let register = u8::try_from(index)
+        let register = u16::try_from(index)
             .map_err(|_| ExecError::InvalidOperand("trampoline argument register"))?;
         process.set_x_reg(register, *arg);
     }
@@ -81,7 +81,7 @@ pub fn handle_trampoline(
         let value = closure.free_var(index).ok_or(ExecError::InvalidOperand(
             "trampoline closure free variable",
         ))?;
-        let register = u8::try_from(usize::from(arity) + index)
+        let register = u16::try_from(usize::from(arity) + index)
             .map_err(|_| ExecError::InvalidOperand("trampoline X register"))?;
         process.set_x_reg(register, value);
     }
